requirements.txt
scipy
pytorch
timm